fix(accordion): resolve toggle functionality issue
