+++ src/components/FAQS.vue
@@ -1,179 +1,130 @@
 <template>
-	<div :class="open ? 'faq open' : 'faq'" @click="$emit('toggleOpen', index)">
-		<div class="accordion-item">
-			<div class="ani">
-				<h2 class="accordion-header" id="headingOne">
-					<button
-						class="accordion-button"
-						:class="{ collapsed: !open }"
-						type="button"
-						aria-expanded="open"
-						aria-controls="collapseOne"
-					>
-						<div class="question">
-							{{ faq.question }}
-						</div>
-					</button>
-				</h2>
-				<div
-					id="collapseOne"
-					class="accordion-collapse"
-					:class="{ collapse: !open, show: open }"
-					aria-labelledby="headingOne"
-					data-parent="#accordionExample"
-				>
-					<div class="accordion-body">
-						<div class="answer" v-html="sanitizeHtml(faq.answer)"></div>
-					</div>
-				</div>
-			</div>
+	<div class="faqs">
+		<h1 class="FAQHEADING">FAQ'S</h1>
+		<div class="faqitems">
+			<FAQItem
+				v-for="(faq, i) in faqs"
+				:faq="faq"
+				:index="i"
+				:key="i"
+				:open="faq.open"
+				@toggleOpen="toggleOpen"
+			/>
 		</div>
 	</div>
 </template>
 
 <script>
-import DOMPurify from 'dompurify'
-
+import FAQItem from './FAQItem.vue'
 export default {
-	name: 'FAQ',
-	props: ['faq', 'index', 'open'],
+	name: 'FAQS',
+	components: {
+		FAQItem,
+	},
+	data() {
+		return {
+			faqs: [
+				{
+					question: 'Why should I participate?',
+
+					answer:
+						'Participating in such events provides an amazing learning experience, helps in improving your development skills, getting involved in big organizations and helping to buildnetworks. This is a great place to get you jump-started and up to speed with various industry-demanded technologies and getting involved with some great Open Source communities.',
+
+					open: false,
+				},
+
+				{
+					question: 'Why is Open Source important?',
+
+					answer:
+						'Open source provides you first-hand experience with working on big andreal-world projects, on which many top companies depend, which will give a huge boost in your development career. Not only this, but you also get to work with and learn from some of the most talented developers around the world.',
+
+					open: false,
+				},
+
+				{
+					question: ' Are there any placement opportunities available?',
+
+					answer:
+						'Contributing to some real world project will have a huge impact on your CV and you will gain some experience about how to work in a large organisation with hundreds of contributors.Though direct placement opportunities are rare, there are chances to get full time opportunities in the organisation you’re contributing for a long time. Mentors often recommend contributors to some startup companies or you might get to know people who work in FAANG companies which can help you later when you apply to those companies with referral.',
+
+					open: false,
+				},
+				{
+					question: 'Can we work in groups to participate in this event?',
+
+					answer:
+						'No, team participation is not allowed in this event. Everyone must participate individually. But you can collaborate with each other if contributing to the same organization.',
+
+					open: false,
+				},
+
+				{
+					question: 'How do I submit my pull requests on Event Website? ',
+
+					answer:
+						'Head over to the Submit PR tab on your dashboard and fill the form accordingly.',
+
+					open: false,
+				},
+
+				{
+					question: 'I am stuck. How do I get help?',
+
+					answer:
+						'Join our discord server (<a href="https://discord.gg/C72FYTYYcZ">click here to join</a>) and the mentors would be glad to help you there.',
+
+					open: false,
+				},
+
+				{
+					question: 'Do I need to pay to enter this event?',
+
+					answer:
+						'No purchase/payment is necessary to participate or win in this event. It is free of cost.',
+
+					open: false,
+				},
+			],
+		}
+	},
+
 	methods: {
-		sanitizeHtml(html) {
-			return DOMPurify.sanitize(html)
+		toggleOpen: function (index) {
+			this.faqs = this.faqs.map((faq, i) => {
+				if (index === i) {
+					faq.open = !faq.open
+				} else {
+					faq.open = false
+				}
+
+				return faq
+			})
 		},
 	},
 }
 </script>
-
-<style scoped>
-* {
-	box-sizing: border-box;
-	background-color: #19192a;
-	color: white;
-}
-.question {
-	color: white;
-	font-size: 23px;
-	width: 80vw;
-}
-
-.answer {
-	color: white;
-}
-
-html,
-body {
-	font-family: 'Arial', sans-serif;
-	font-size: 19px;
-	line-height: 1.5;
-	color: white;
-	width: 100%;
-	height: 100%;
-	background-color: #19192a;
+<style>
+.FAQHEADING {
+	font-family: Stargaze, serif;
+	font-size: 6vw;
 	margin: 0;
-	padding: 0;
-}
-
-body {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	flex-direction: column;
-}
-
-h1 {
-	font-family: 'Nexa Black', sans-serif;
-	font-weight: 700;
-	font-size: 1.8rem;
-	margin-bottom: 30px;
-}
-
-.accordion {
-	width: 500px;
-	box-shadow: 0 2px 5px rgba(0, 0, 0, 0.25);
-}
-
-.ani {
-	border-bottom: 1px white solid;
-	width: 50vw;
-}
-.accordion-item {
-	display: flex;
-	justify-content: center;
-	width: 100vw;
-}
-
-.accordion-button {
-	display: flex;
-	align-items: center;
-
-	padding: 10px;
+	text-align: center;
+	padding-top: 7vh;
+	color: #fffaff;
+	padding-bottom: 60px;
 	background-color: #19192a;
-	border: none;
-	text-align: left;
-	transition: background-color 0.4s ease;
-	font-family: 'Nexa XBold', sans-serif;
-	font-size: 1rem;
-	color: white;
-	position: relative;
+	padding-top: 30px;
+	padding-bottom: 50px;
 }
-
-.accordion-button.collapsed::after,
-.accordion-button:not(.collapsed)::after {
-	content: ' ';
-	width: 30px;
-	height: 26px;
-	background-repeat: no-repeat;
-	background-image: url('../assets/arrow-204-32.png');
-	position: absolute;
-	right: 0px;
-	transform: scale(0.7);
-	transition: all ease 0.4s;
-}
-
-.accordion-button:not(.collapsed)::after {
-	transform: rotate(180deg) scale(0.7);
-}
-
-.accordion-collapse {
-	border: none;
-	max-height: 0;
-	overflow: hidden;
-	transition: max-height 0.4s ease;
-}
-
-.show {
-	max-height: 1000px;
-	animation: expandAccordion ease 3s;
-}
-
-.accordion-body {
-	padding: 15px 10px;
+.faqitems {
+	padding-bottom: 100px;
 	background-color: #19192a;
 }
 
-.answer {
-	font-size: 0.9rem;
-	color: white;
-}
-
-@keyframes expandAccordion {
-	from {
-		max-height: 0;
-	}
-	to {
-		max-height: 1000px;
-	}
-}
-@media (max-width: 520px) {
-	.ani {
-		width: 90%;
-	}
-}
-
-@media (max-width: 600px) and (min-width: 520px) {
-	.ani {
-		width: 68vw;
+@media (max-width: 500px) {
+	.FAQHEADING {
+		font-size: 15vw;
 	}
 }
 </style>
